refactor(file conflicts): Refactor types into file conflict file PE-639
