+++ src/types.ts
@@ -9,6 +9,12 @@ export const upsertOnConflicts = 'upsert';
 export const askOnConflicts = 'ask';
 export const renameOnConflicts = 'rename';
 
+export type FileNameConflictResolution =
+	| typeof skipOnConflicts
+	| typeof replaceOnConflicts
+	| typeof upsertOnConflicts
+	| typeof askOnConflicts;
+
 export const JSON_CONTENT_TYPE = 'application/json';
 export const PRIVATE_CONTENT_TYPE = 'application/octet-stream';
